Component files update
